pyproject.toml
[tool.poetry]
name = "strawberry-django-export-schema-example"
version = "0.1.0"
description = ""
authors = ["Joe Cox <joe@medallion.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "3.2.20"
strawberry-graphql = "0.209.1"
strawberry-graphql-django = "0.17.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
